move image to own pkg
